create search component

+++ components/Main.jsx
@@ -1,12 +1,22 @@
 import {useEffect, useState} from 'react';
-import {View, Text, TouchableOpacity} from 'react-native';
-import {AnimatedFAB, Modal, Portal, Snackbar} from 'react-native-paper';
+import {View, Text, TouchableOpacity, Image, Linking} from 'react-native';
+import {
+  AnimatedFAB,
+  Button,
+  Divider,
+  Menu,
+  Modal,
+  Portal,
+  Snackbar,
+} from 'react-native-paper';
 import AddWord from './AddWord';
 import NotWord from './NotWord';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import TableWord from './TableWord';
 import ExamWord from './ExamWord';
 import EditWord from './EditWord';
+import SearchWord from './SearchWord';
+import AboutMe from './AboutMe';
 
 const Main = () => {
   const [error, setError] = useState({show: false, msg: ''});
@@ -17,18 +27,27 @@ const Main = () => {
     persian: '',
   });
   const [visible, setVisible] = useState(false);
+  const [menuVisible, setMenuVisible] = useState(false);
+  const [searchWordVisible, setSearchWordVisible] = useState(false);
+  const [aboutMeVisible, setAboutMeVisible] = useState(false);
   const [page, setPage] = useState('noword');
   const [words, setWords] = useState([]);
   const [test, setTest] = useState(false);
   const [testWord, setTestWord] = useState({});
   const [currentWord, setCurrentWord] = useState({});
- const [screen , setScreen] = useState(false)
+  const [searchWord, setSearchWord] = useState({});
+  const [screen, setScreen] = useState(false);
   const [currentWordVisible, setCurrentWordVisible] = useState(false);
 
   const hideModal = () => setVisible(false);
   const hideTestModal = () => setTest(false);
   const hideCurrentWordVisible = () => setCurrentWordVisible(false);
-
+  const closeMenu = () => setMenuVisible(false);
+  const hideAboutMeVisible = () => setAboutMeVisible(false);
+  const hideSearchWordVisible = () => {
+    setSearchWord({});
+    setSearchWordVisible(false);
+  };
   const rebuildHandle = () => {
     const range = Math.ceil(Math.random() * words.length - 1);
     setTestWord(words[range]);
@@ -37,7 +56,14 @@ const Main = () => {
   const addHandler = async () => {
     if (textValue.english === '' || textValue.persian === '') return;
 
-    if (Boolean(words.find(english => english.english == textValue.english))) {
+    if (
+      Boolean(
+        words.find(
+          english =>
+            english.english.toLowerCase() == textValue.english.toLowerCase(),
+        ),
+      )
+    ) {
       setError({show: true, msg: 'This word duplicate.'});
       setTextValue({
         id: '',
@@ -78,6 +104,22 @@ const Main = () => {
     });
   };
 
+  const resetHandler = () => {
+    setSearchWord({});
+  };
+  const searchHandler = word => {
+    let w = /[پچجحخهعغفقثصضشسیبلاتنمکگوئدذرزطظژؤإأءًٌٍَُِّ\s]+$/;
+    if (word.match(w)) {
+      const pWord = words.find(item => item.persian == word);
+      setSearchWord(pWord);
+    } else {
+      const eWord = words.find(
+        item => item.english.toLowerCase() == word.toLowerCase(),
+      );
+      setSearchWord(eWord);
+    }
+  };
+
   const deleteHandle = async e => {
     const newWordList = words.filter(item => item.id != e);
     setWords(newWordList);
@@ -91,13 +133,21 @@ const Main = () => {
   };
 
   const changeWord = async (oldWord, newWord) => {
-    const result = words.map((item, index) =>
-      item.id == oldWord.id ? (item[index] = newWord) : item,
+    const findNewWordInWords = words.find(
+      item => item.english.toLowerCase() === newWord.english.toLowerCase(),
     );
-    setWords(result);
-    await AsyncStorage.clear()
-    await AsyncStorage.setItem('words', JSON.stringify(result));
-    setCurrentWordVisible(false);
+    if (!findNewWordInWords) {
+      const result = words.map((item, index) =>
+        item.id == oldWord.id ? (item[index] = newWord) : item,
+      );
+      setWords(result);
+      await AsyncStorage.clear();
+      await AsyncStorage.setItem('words', JSON.stringify(result));
+      setCurrentWordVisible(false);
+    } else {
+      setError({show: true, msg: 'This word duplicate.'});
+      return;
+    }
   };
 
   useEffect(() => {
@@ -131,19 +181,56 @@ const Main = () => {
         </View>
       ) : (
         <View style={{flex: 1}}>
-          <View
-            style={{flex: 1}}
-            className="flex flex-row items-center justify-between p-3 bg-black">
-            <Text className="text-white">Words: {words.length}</Text>
+          <View style={{flex: 1}} className="w-full py-3 bg-black">
+            <View className="w-[95%] h-full mx-auto flex flex-row items-center justify-between">
+              <Text className="text-white pl-5">Words: {words.length}</Text>
 
-            <TouchableOpacity
-              onPress={() => {
-                setTest(true);
-                rebuildHandle();
-              }}
-              className="bg-red-300 py-2 px-6 rounded-md">
-              <Text>Test</Text>
-            </TouchableOpacity>
+              <View>
+                <Menu
+                  visible={menuVisible}
+                  onDismiss={closeMenu}
+                  anchorPosition={'bottom'}
+                  anchor={
+                    <TouchableOpacity
+                      onPress={() => setMenuVisible(true)}
+                      className=" flex justify-center">
+                      <Button
+                        icon="menu"
+                        labelStyle={{fontSize: 24, color: 'white'}}
+                      />
+                    </TouchableOpacity>
+                  }>
+                  <Menu.Item
+                    leadingIcon="search-web"
+                    onPress={() => {
+                      closeMenu();
+                      setSearchWordVisible(true);
+                      rebuildHandle();
+                    }}
+                    title="Search word"
+                  />
+
+                  <Menu.Item
+                    leadingIcon="test-tube"
+                    onPress={() => {
+                      closeMenu();
+                      setTest(true);
+                      rebuildHandle();
+                    }}
+                    title="Exam"
+                  />
+                  <Divider />
+                  <Menu.Item
+                    leadingIcon="alert-decagram"
+                    onPress={() => {
+                      closeMenu();
+                      setAboutMeVisible(true);
+                    }}
+                    title="About me"
+                  />
+                </Menu>
+              </View>
+            </View>
           </View>
           <View style={{flex: 10}}>
             <TableWord
@@ -157,14 +244,13 @@ const Main = () => {
         </View>
       )}
       {!screen && (
-
         <AnimatedFAB
-        icon="plus"
-        label={'Label'}
-        onPress={() => setVisible(true)}
-        className="absolute bottom-5 right-5"
+          icon="plus"
+          label={'Label'}
+          onPress={() => setVisible(true)}
+          className="absolute bottom-5 right-5"
         />
-        )}
+      )}
 
       <Portal>
         <Modal
@@ -206,6 +292,36 @@ const Main = () => {
             marginHorizontal: 20,
           }}>
           <EditWord word={currentWord} changeWord={changeWord} />
+        </Modal>
+      </Portal>
+
+      <Portal>
+        <Modal
+          visible={searchWordVisible}
+          onDismiss={hideSearchWordVisible}
+          contentContainerStyle={{
+            backgroundColor: 'white',
+            padding: 20,
+            marginHorizontal: 20,
+          }}>
+          <SearchWord
+            searchWordObject={searchWord}
+            searchHandler={searchHandler}
+            resetHandler={resetHandler}
+          />
+        </Modal>
+      </Portal>
+
+      <Portal>
+        <Modal
+          visible={aboutMeVisible}
+          onDismiss={hideAboutMeVisible}
+          contentContainerStyle={{
+            backgroundColor: 'white',
+            padding: 20,
+            marginHorizontal: 20,
+          }}>
+          <AboutMe />
         </Modal>
       </Portal>
